% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.dummy.R
\name{as.dummy}
\alias{as.dummy}
\title{as.dummy}
\usage{
as.dummy(x)
}
\arguments{
\item{x}{a vector of categorical values}
}
\value{
a matrix of dummy variable, where each column represents a category.
}
\description{
as.dummy transforms a categorical variable to a dummy variable.
}
\examples{
x <- rep( c('red','yellow','blue') , each=10 )
dx <- as.dummy(x)
}
