% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixEpistasis_main.R
\name{MatrixEpistasis_main}
\alias{MatrixEpistasis_main}
\title{matrixEpistasis}
\usage{
MatrixEpistasis_main(snpA, snpB, trait, covariate = NULL,
  pvalThreshold = 1e-05, outputFileName)
}
\arguments{
\item{snpA}{a matrix of numeric values in size of sample*snp representing the 1st group of SNPs, where the column names are the snp_ids}

\item{snpB}{a matrix of numeric values in size of sample*snp representing the 2nd group of SNPs, where the column names are the snp_ids}

\item{trait}{a vector of numeric values representing the quantitative trait}

\item{covariate}{a matrix of numeric values in size of sample*covariate, by default, NULL}

\item{pvalThreshold}{a numeric value representing the p-value threshold for matrixEpistasis output}

\item{outputFileName}{a character value indicating the file name of matrixEpistasis output}
}
\value{

}
\description{
matrixEpistasis uses large matrix operation to perform the exhaustive epistasis scan for quantitative traits with covariate adjustment
}
\examples{
# randomly generate a SNP matrix
snp <- sapply(1:100,function(i) rnorm(1000) )
# assign names to SNPs
colnames(snp) <- paste0('snp',1:100)
snpA = snp
snpB = snp
# radnomly generate a quantitative trait by simulating the relationship between SNPs and traits
trait <- snp \%*\% rnorm(100)
# use the top 5 PCs as the covariates 
covariate <- prcomp(snp)$x[,1:5]

MatrixEpistasis_main( snpA=snpA, snpB=snpB, trait=trait, covariate=covariate, pvalThreshold=1e-5, outputFileName='matrixEpistasis_pval_1e-5.txt')



}
\references{

}
\seealso{
\code{\link{matrixEpistasis}}; \code{\link{matrixPval}}; \code{\link{p2c}};
}
\author{
Shijia Zhu, \email{shijia.zhu@mssm.edu}
}
