% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPar.R
\name{matrixPar}
\alias{matrixPar}
\title{matrixPar}
\usage{
matrixPar(r12, r13, r23)
}
\arguments{
\item{r12}{a matrix of correlation coefficients between the 1st and 2nd groups of varaibles}

\item{r13}{a matrix of correlation coefficients between the 1st and 3rd groups of varaibles}

\item{r23}{a matrix of correlation coefficients between the 2nd and 3rd groups of varaibles}
}
\value{
a matrix of partial correlation coefficients between the 1st and 2nd groups of variables conditioned on the 3rd
}
\description{
matrixPar calculates the partial correlation using the iterative formula
}
\examples{

}
