% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixQtl.R
\name{matrixQtl}
\alias{matrixQtl}
\title{matrixQtl}
\usage{
matrixQtl(snp, trait, covariate = NULL)
}
\arguments{
\item{snp}{a matrix of numeric values in size of sample*snp}

\item{trait}{a matrix of numeric values in size of sample*trait}

\item{covariate}{a matrix of numeric values  in size of sample*covariate}
}
\value{
A list containing two components:
\itemize{
\item{r} a matrix of Pearson correlation between snp and trait in size of snp*trait
\item{df} the degree of freedom
}
}
\description{
matrixQtl calculates the correlation between snp and quantitative trait
}
\examples{
snp <- sapply(1:100,function(i) rnorm(1000) )
alpha <- sapply(1:10,function(i) rnorm(100) )
trait <- snp \%*\% alpha
covariates <- prcomp(snp)$x[,1:5]

qtl <- matrixQtl( snp, trait , covariates )
pval1 <- matrixPval( r=qtl$r , df=qtl$df)

pval2 <- matrix(nrow=ncol(snp),ncol=ncol(trait),data=0)
for(i in 1:ncol(snp))
{
  for(j in 1:ncol(trait))
  {
    pval2[i,j] <- summary(lm(trait[,j]~snp[,i]+covariates))$coef[2,4]
  }
}


head(pval1)
head(pval2)

}
\references{

}
\seealso{
\code{\link{matrixPval}};
}
\author{
Shijia Zhu, \email{shijia.zhu@mssm.edu}
}
