% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2c.R
\name{p2c}
\alias{p2c}
\title{p2c}
\usage{
p2c(pval, df)
}
\arguments{
\item{pval}{a numeric value in [0,1] representing the p value}

\item{df}{an integer value representing the degree of freedom}
}
\value{
a numeric value in [0,1] representing the absolute value of correlation coefficient
}
\description{
p2c converts the given p value to the absolute value of corresponding correlation coefficient
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100)
z <- cor.test(x,y)
z
p2c(z$p.value,98)

}
\author{
Shijia Zhu, \email{shijia.zhu@mssm.edu}
}
