% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNAN.R
\name{removeNAN}
\alias{removeNAN}
\title{removeNAN}
\usage{
removeNAN(x, v = 0)
}
\arguments{
\item{x}{a matrix of numeric values}

\item{v}{the value assigned to the missing values, by default, 0}
}
\value{
a matrix of numeric values
}
\description{
removeNAN removes the missing values from the data matrix.
}
\examples{
x <- matrix(c(0,1,NA,2),ncol=2)
x
removeNAN(x)
}
