% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothQQplot.R
\name{smoothQQplot}
\alias{smoothQQplot}
\title{smoothQQplot}
\usage{
smoothQQplot(x, y)
}
\arguments{
\item{x}{a vector of numeric values in [0,1] representing the first group of p values}

\item{y}{a vector of numeric values in [0,1] representing the second group of p values}
}
\value{

}
\description{
smoothQQplot draws the QQplot between two groups of p values together with the smoothed density representation of the scatterplot
}
\examples{
x <- runif(10000)
y <- runif(10000)
smoothQQplot(x,y)
}
